renamed pokemonHandler to pokemon
pokemonHandler is a function while the file contains lot more functions related to handling the pokemon
